refactor(agent): extract residual module-scope helpers out of dkg-agent.ts
Phase 0 of the dkg-agent.ts size reduction. Moves the remaining
module-private constants, free functions, and the SwmSenderKeySetupRejectionError
class that still lived above the DKGAgent class into the existing sibling
modules, following the established extraction convention:

- raceWithBootTimeout / isTransientBootChainError -> new dkg-agent-boot.ts
- sliceIntoCiphertextChunks -> dkg-agent-helpers.ts
- SwmSenderKeySetupRejectionError -> dkg-agent-types.ts
- boot/chunk/policy timeout constants + sender-key drain log ctx -> dkg-agent-constants.ts

Pure 1:1 relocations, no behavior change. Also drops the now-unused
isRetryableRpcError import. Typecheck clean; agent unit lane (123 tests) green.
